put store logic in store itself
